langchain-agent-tutorial/requirements.txt
# LangChain 核心库
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LangGraph - 用于构建智能体工作流
langgraph>=0.2.0

# 阿里千问模型支持
dashscope>=1.20.0

# OpenAI 兼容接口 (千问支持 OpenAI 兼容 API)
openai>=1.0.0

# 其他工具
python-dotenv>=1.0.0
httpx>=0.25.0
